requirements.txt
Django==1.11.20
psycopg2>=2.7,<3.0
djangorestframework==3.6.3